Move helpers to `dom-utils.ts`
