+++ source/helpers/dom-utils.ts
@@ -66,3 +66,13 @@ export const removeClassFromAll = (className: string): void => {
 export const frame = async (): Promise<number> => new Promise(resolve => {
 	requestAnimationFrame(resolve);
 });
+
+export const highlightTab = (tabElement: Element): void => {
+	tabElement.classList.add('selected');
+	tabElement.setAttribute('aria-current', 'page');
+};
+
+export const unhighlightTab = (tabElement: Element): void => {
+	tabElement.classList.remove('selected');
+	tabElement.removeAttribute('aria-current');
+};
